Update update_users_active_state.rb not to autorunning production environment
